[Foundation] Move the InflightData out of the handler file.
