+++ src/Foundation/NSUrlSessionHandler.cs
@@ -942,65 +942,5 @@ namespace Foundation {
 				return true;
 			}
 		}
-
-		class InflightData : IDisposable
-		{
-			public readonly object Lock = new object ();
-			public string RequestUrl { get; set; }
-
-			public TaskCompletionSource<HttpResponseMessage> CompletionSource { get; } = new TaskCompletionSource<HttpResponseMessage> ();
-			public CancellationToken CancellationToken { get; set; }
-			public CancellationTokenSource CancellationTokenSource { get; } = new CancellationTokenSource ();
-			public NSUrlSessionDataTaskStream Stream { get; } = new NSUrlSessionDataTaskStream ();
-			public HttpRequestMessage Request { get; set; }
-			public HttpResponseMessage? Response { get; set; }
-
-			public Exception? Exception { get; set; }
-			public bool ResponseSent { get; set; }
-			public bool Errored { get; set; }
-			public bool Disposed { get; set; }
-			public bool Completed { get; set; }
-			public bool Done { get { return Errored || Disposed || Completed || CancellationToken.IsCancellationRequested; } }
-
-			public InflightData (string requestUrl, CancellationToken cancellationToken, HttpRequestMessage request)
-			{
-				RequestUrl = requestUrl;
-				CancellationToken = cancellationToken;
-				Request = request;
-			}
-
-			public void Dispose()
-			{
-				Dispose (true);
-				GC.SuppressFinalize(this);
-			}
-
-			// The bulk of the clean-up code is implemented in Dispose(bool)
-			protected virtual void Dispose (bool disposing)
-			{
-				if (disposing) {
-					CancellationTokenSource.Dispose ();
-				}
-			}
-
-		}
-
-		class NSUrlSessionDataTaskStreamContent : MonoStreamContent
-		{
-			Action? disposed;
-
-			public NSUrlSessionDataTaskStreamContent (NSUrlSessionDataTaskStream source, Action onDisposed, CancellationToken token) : base (source, token)
-			{
-				disposed = onDisposed;
-			}
-
-			protected override void Dispose (bool disposing)
-			{
-				var action = Interlocked.Exchange (ref disposed, null);
-				action?.Invoke ();
-
-				base.Dispose (disposing);
-			}
-		}
 	}
 }
