+++ src/Makefile
@@ -156,6 +156,7 @@ IOS_DOTNET_HTTP_SOURCES = \
 	Foundation/NSUrlSessionHandler/NSHttpCookieExtensions.cs \
 	Foundation/NSUrlSessionHandler/NSUrlSessionDataTaskStream.cs \
 	Foundation/NSUrlSessionHandler/NSUrlSessionDataTaskStreamContent.cs \
+	Foundation/NSUrlSessionHandler/InflightData.cs \
 	Foundation/NSUrlSessionHandler/MonoStreamContent.cs \
 	Foundation/NSUrlSessionHandler/WrappedNSInputStream.cs \
 	System.Net.Http/CFContentStream.cs \
@@ -166,6 +167,7 @@ IOS_HTTP_SOURCES = \
 	Foundation/NSUrlSessionHandler/NSHttpCookieExtensions.cs \
 	Foundation/NSUrlSessionHandler/NSUrlSessionDataTaskStream.cs \
 	Foundation/NSUrlSessionHandler/NSUrlSessionDataTaskStreamContent.cs \
+	Foundation/NSUrlSessionHandler/InflightData.cs \
 	Foundation/NSUrlSessionHandler/MonoStreamContent.cs \
 	Foundation/NSUrlSessionHandler/WrappedNSInputStream.cs \
 	System.Net.Http/CFContentStream.cs \
@@ -549,6 +551,7 @@ MAC_HTTP_SOURCES = \
 	Foundation/NSUrlSessionHandler/NSHttpCookieExtensions.cs \
 	Foundation/NSUrlSessionHandler/NSUrlSessionDataTaskStream.cs \
 	Foundation/NSUrlSessionHandler/NSUrlSessionDataTaskStreamContent.cs \
+	Foundation/NSUrlSessionHandler/InflightData.cs \
 	Foundation/NSUrlSessionHandler/MonoStreamContent.cs \
 	Foundation/NSUrlSessionHandler/WrappedNSInputStream.cs \
 	System.Net.Http/CFContentStream.cs \
@@ -869,6 +872,7 @@ WATCHOS_HTTP_SOURCES = \
 	Foundation/NSUrlSessionHandler/NSHttpCookieExtensions.cs \
 	Foundation/NSUrlSessionHandler/NSUrlSessionDataTaskStream.cs \
 	Foundation/NSUrlSessionHandler/NSUrlSessionDataTaskStreamContent.cs \
+	Foundation/NSUrlSessionHandler/InflightData.cs \
 	Foundation/NSUrlSessionHandler/MonoStreamContent.cs \
 	Foundation/NSUrlSessionHandler/WrappedNSInputStream.cs \
 
@@ -1166,6 +1170,7 @@ TVOS_DOTNET_HTTP_SOURCES = \
 	Foundation/NSUrlSessionHandler/NSHttpCookieExtensions.cs \
 	Foundation/NSUrlSessionHandler/NSUrlSessionDataTaskStream.cs \
 	Foundation/NSUrlSessionHandler/NSUrlSessionDataTaskStreamContent.cs \
+	Foundation/NSUrlSessionHandler/InflightData.cs \
 	Foundation/NSUrlSessionHandler/MonoStreamContent.cs \
 	Foundation/NSUrlSessionHandler/WrappedNSInputStream.cs \
 	System.Net.Http/CFContentStream.cs \
@@ -1177,6 +1182,7 @@ TVOS_HTTP_SOURCES = \
 	Foundation/NSUrlSessionHandler/NSHttpCookieExtensions.cs \
 	Foundation/NSUrlSessionHandler/NSUrlSessionDataTaskStream.cs \
 	Foundation/NSUrlSessionHandler/NSUrlSessionDataTaskStreamContent.cs \
+	Foundation/NSUrlSessionHandler/InflightData.cs \
 	Foundation/NSUrlSessionHandler/MonoStreamContent.cs \
 	Foundation/NSUrlSessionHandler/WrappedNSInputStream.cs \
 	System.Net.Http/CFContentStream.cs \
@@ -1395,6 +1401,7 @@ MACCATALYST_DOTNET_HTTP_SOURCES = \
 	Foundation/NSUrlSessionHandler/NSHttpCookieExtensions.cs \
 	Foundation/NSUrlSessionHandler/NSUrlSessionDataTaskStream.cs \
 	Foundation/NSUrlSessionHandler/NSUrlSessionDataTaskStreamContent.cs \
+	Foundation/NSUrlSessionHandler/InflightData.cs \
 	Foundation/NSUrlSessionHandler/MonoStreamContent.cs \
 	Foundation/NSUrlSessionHandler/WrappedNSInputStream.cs \
 	System.Net.Http/CFContentStream.cs \
@@ -1406,6 +1413,7 @@ MACCATALYST_HTTP_SOURCES = \
 	Foundation/NSUrlSessionHandler/NSHttpCookieExtensions.cs \
 	Foundation/NSUrlSessionHandler/NSUrlSessionDataTaskStream.cs \
 	Foundation/NSUrlSessionHandler/NSUrlSessionDataTaskStreamContent.cs \
+	Foundation/NSUrlSessionHandler/InflightData.cs \
 	Foundation/NSUrlSessionHandler/MonoStreamContent.cs \
 	Foundation/NSUrlSessionHandler/WrappedNSInputStream.cs \
 	System.Net.Http/CFContentStream.cs \
